cargo sort, forge fmt

+++ contracts/DataFeedFeeder.sol
@@ -36,15 +36,18 @@ contract DataFeedFeeder {
     ///         (in this case, checking if a number is even) are considered valid.
 
     bytes32 public constant imageId = ImageID.JSON_PARSER_ID;
-    mapping (string => DataFeedStorage) dataFeedStorages;
-    uint constant PAIRS_AMOUNT = 4;
+    mapping(string => DataFeedStorage) dataFeedStorages;
+    uint256 constant PAIRS_AMOUNT = 4;
 
     /// @notice Initialize the contract, binding it to a specified RISC Zero verifier.
-    constructor(IRiscZeroVerifier _verifier, string[PAIRS_AMOUNT] memory /* array size must be fixed in memory */ pair_names) {
+    constructor(
+        IRiscZeroVerifier _verifier,
+        string[PAIRS_AMOUNT] memory /* array size must be fixed in memory */ pair_names
+    ) {
         verifier = _verifier;
 
-        for (uint i = 0; i < pair_names.length; i++) {
-            dataFeedStorages[pair_names[i]] = new DataFeedStorage(pair_names[i], 5 /* TODO hardcoded*/);
+        for (uint256 i = 0; i < pair_names.length; i++) {
+            dataFeedStorages[pair_names[i]] = new DataFeedStorage(pair_names[i], 5 /* TODO hardcoded*/ );
         }
     }
 
@@ -58,14 +61,14 @@ contract DataFeedFeeder {
     ) public {
         // Construct the expected journal data. Verify will fail if journal does not match.
         PairDataStruct[PAIRS_AMOUNT] memory pairs_data;
-        for (uint i = 0; i < PAIRS_AMOUNT; i++) {
+        for (uint256 i = 0; i < PAIRS_AMOUNT; i++) {
             pairs_data[i] = PairDataStruct(pair_names[i], prices[i], timestamps[i]);
         }
         bytes memory journal = abi.encode(pairs_data);
         verifier.verify(seal, imageId, sha256(journal));
 
-        for (uint i = 0; i < pairs_data.length; i++) {
-            PairDataStruct memory  _current_pair_data = pairs_data[i];
+        for (uint256 i = 0; i < pairs_data.length; i++) {
+            PairDataStruct memory _current_pair_data = pairs_data[i];
             uint64 _price = _current_pair_data.price;
             uint64 _timestamp = _current_pair_data.timestamp;
             string memory _pair_name = _current_pair_data.pair_name;
